// app/src/services/spacefinder-sdk/model/Booking.ts
/**
 * Spacefinder-API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Booking {
    bookingId?: string;

    locationId: string;

    locationName?: string;

    resourceId: string;

    userId: string;

    userFirstName?: string;

    userLastName?: string;

    resourceName?: string;

    resourceDescription?: string;

    timeslot?: string;

    startTimeIsoTimestamp?: string;

    startTimeEpochTime?: number;

    endTimeIsoTimestamp?: string;

    endTimeEpochTime?: number;

}